requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# Feature Matching Models
kornia>=0.7.0
kornia-rs>=0.1.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# Image Processing
pillow>=10.0.0
albumentations>=1.3.0

# CLIP for fallback
ftfy>=6.1.0
regex>=2023.0.0

# Tracking & Filtering
filterpy>=1.4.5

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
einops>=0.7.0

# Optional: For visualization
matplotlib>=3.7.0
seaborn>=0.12.0

